【AC】[AtCoder Beginner Contest 044]C - Tak and Cards

+++ abc044/arc060_a.cpp
@@ -29,14 +29,14 @@ typedef priority_queue<ll, vector<ll>, greater<ll>> PQ_ASK;
 const int mod = 1000000007;
 
 
-void knapsack(vector<int> &numbers, vector<int> &ans) {
+void knapsack(vector<ll> &numbers, vector<ll> &ans) {
 
 
-    vector<vector<int>> dp(60, vector<int>(60 * 60, 0));
+    vector<vector<ll>> dp(60, vector<ll>(60 * 60, 0));
     dp[0][0] = 1;
 
-    for (int i = 0; i < numbers.size(); ++i) {
-        for (int j = 0; j < 60 * 60; ++j) {
+    for (ll i = 0; i < numbers.size(); ++i) {
+        for (ll j = 0; j < 60 * 60; ++j) {
             dp[i + 1][j] += dp[i][j];
             if (j >= numbers[i]) {
                 dp[i + 1][j] += dp[i][j - numbers[i]];
@@ -44,7 +44,7 @@ void knapsack(vector<int> &numbers, vector<int> &ans) {
         }
     }
 
-    for (int i = 0; i < ans.size(); i++) {
+    for (ll i = 0; i < ans.size(); i++) {
         ans[i] = dp[numbers.size()][i];
     }
 
@@ -53,13 +53,13 @@ void knapsack(vector<int> &numbers, vector<int> &ans) {
 
 int main() {
 
-    int n, a;
+    ll n, a;
     cin >> n >> a;
-    vector<int> plus;
-    vector<int> minus;
-    int equal = 0;
+    vector<ll> plus;
+    vector<ll> minus;
+    ll equal = 0;
     rep(i, n) {
-        int x;
+        ll x;
         cin >> x;
         if (x > a) {
             plus.push_back(x - a);
@@ -70,8 +70,8 @@ int main() {
         }
     }
 //
-    vector<int> ps(60 * 60);
-    vector<int> ms(60 * 60);
+    vector<ll> ps(60 * 60);
+    vector<ll> ms(60 * 60);
 //    knapsack(plus);
     knapsack(plus, ps);
 //    knapsack(minus);
@@ -79,16 +79,16 @@ int main() {
 
 
     ll ans = -1;
-    for (int i = 0; i < 60 * 60; i++) {
+    for (ll i = 0; i < 60 * 60; i++) {
         ans += ps[i] * ms[i];
     }
 
-    for (int k = 0; k < equal; k++) {
+    for (ll k = 0; k < equal; k++) {
         ans *= 2;
     }
 
     ll eq_ans = 1;
-    for (int k = 0; k < equal; k++) {
+    for (ll k = 0; k < equal; k++) {
         eq_ans *= 2;
     }
